// src/enums/Category.ts
export enum Category {
  Tutorial = "Tutorial",
  Triple = "Triple",
  Quintuple = "Quintuple",

  Battle = "Battle",
  Challenge = "Challenge",
  Cranky = "Cranky",
  Rehearsal = "Rehearsal",
  ShiftySmash = "ShiftySmash",

  Bloomintone = "Bloomintone",
  Chiptone = "Chiptone",
  Fragmentone = "Fragmentone",
  Glamtone = "Glamtone",
  Growtone = "Growtone",
  Quiztone = "Quiztone",
  Seekertone = "Seekertone",
  Smashintone = "Smashintone"
}